# resources/openapi.yaml
openapi: 3.0.0
info:
  title: Refinery Procurement System API
  description: API for managing refinery equipment catalog and procurement orders.
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
paths:
...
  /admin/audit/user/{id}:
    get:
      summary: Get audit logs by user
      description: Retrieve audit logs for a specific user.
      parameters:
        - in: path
          name: id
          schema:
            type: string
            format: uuid
          required: true
          description: User UUID
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AuditLogResponse'
  /admin/audit/target:
    get:
      summary: Get audit logs by target
      description: Retrieve audit logs for a specific target entity.
      parameters:
        - in: query
          name: entity
          schema:
            type: string
          required: true
          description: Target entity type (e.g., "purchase_order")
        - in: query
          name: id
          schema:
            type: string
          required: true
          description: Target entity ID
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AuditLogResponse'
components:
  schemas:
...
    AuditLogResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        user_id:
          type: string
          format: uuid
        action:
          type: string
        target_entity:
          type: string
        target_id:
          type: string
        changes:
          type: object
          nullable: true
        created_at:
          type: string
          format: date-time